I've greatly simplified the questionnaire controller by storing the current decision state in the page as hidden form fields rather than in the session cookie. To do this I have removed the Redirect-on-post handling and am rendering the question and results pages directly from the post controller. The downside is that if the user uses the back button or refreshes then they may see a repost-this-page popup, but because no data is being stored all the requests are still idempotent.
