+++ src/main/scala/models/DecisionState.scala
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2017  Department for Business, Energy and Industrial Strategy
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package models
+
+import questionnaire.{FinancialYear, Thresholds}
+import utils.YesNo
+
+case class DecisionState(
+                          isCompanyOrLLP: Option[YesNo],
+                          financialYear: Option[FinancialYear],
+                          companyThresholds: Thresholds,
+                          subsidiaries: Option[YesNo],
+                          subsidiaryThresholds: Thresholds
+                        )
+
+object DecisionState {
+  val empty: DecisionState = models.DecisionState(None, None, Thresholds.empty, None, Thresholds.empty)
+
+  import YesNo.Yes
+
+  val secondYear: DecisionState = models.DecisionState(Some(Yes), Some(FinancialYear.Second), Thresholds(Some(Yes), Some(Yes), Some(Yes)), Some(Yes), Thresholds(Some(Yes), Some(Yes), Some(Yes)))
+  val thirdYear: DecisionState = models.DecisionState(Some(Yes), Some(FinancialYear.ThirdOrLater), Thresholds(Some(Yes), Some(Yes), Some(Yes)), Some(Yes), Thresholds(Some(Yes), Some(Yes), Some(Yes)))
+}
