+++ src/main/scala/questionnaire/Summarizer.scala
@@ -19,6 +19,7 @@ package questionnaire
 
 import javax.inject.Inject
 
+import models.DecisionState
 import play.api.i18n.MessagesApi
 import utils.YesNo
