+++ src/main/twirl/views/questionnaire/start.scala.html
@@ -13,5 +13,5 @@
         Your answers must be for your business on its own, even if it's part of a group.
         Every business within the group will need to do this individually.
     </p>
-    <a name="Start now" href="@routes.QuestionnaireController.startQuestions()" class="button button-start">Start now</a>
+    <a name="Start now" href="@routes.QuestionnaireController.nextQuestion()" class="button button-start">Start now</a>
 </div>
